fix(xilinx): resolve misra rule 10.3 violations
Fixed below MISRA violation:
- MISRA Violation: MISRA-C:2012 R.10.3:
  - The value of an expression shall not be assigned to an object
    with a narrower essential type or of a different essential type
    category.
- Fix:
  - Explicitly type casted to narrower essential type or of a different
    essential type category.

+++ plat/xilinx/versal/sip_svc_setup.c
@@ -1,6 +1,6 @@
 /*
  * Copyright (c) 2018-2021, Arm Limited and Contributors. All rights reserved.
- * Copyright (c) 2022-2024, Advanced Micro Devices, Inc. All rights reserved.
+ * Copyright (c) 2022-2025, Advanced Micro Devices, Inc. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -116,6 +116,6 @@ DECLARE_RT_SVC(
 		sip_svc,
 		OEN_SIP_START,
 		OEN_SIP_END,
-		SMC_TYPE_FAST,
+		(uint8_t)SMC_TYPE_FAST,
 		sip_svc_setup,
 		sip_svc_smc_handler);
